:building_construction: Added auth controller
